Inject our license widget to a widget sidebar when switching themes.
By default, the post license is displayed at the end of a post if the
the license widget is not displayed anywhere in the theme and if the
post license differs from the site license.

For those that might forget to add our license widget in their theme,
we are now going to inject our license widget into one of the theme's
sidebars when the theme changes.

Note: When attempting to inject, we check if one of the theme's
sidebars contains either 'footer' or 'side' in the sidebar ID to
ensure our widget will display in the footer or sidebar.  Preference
is given to the footer.